Refactor SourcePlan existence query into fxspot sourcing module
